#11 - fix extraProperties issue
Not sure of the ideal behaviour here:

The bug was being caused by:

The getter/setter being defined on the first instance to create a property at runtime, but the definition of the property being stored in this._definition, which is the same object as prototype._definition. Thus, when the second model tries to add the property, the definition is already in the prototype hash, so it doesn’t go ahead and create the getter/setter on the instance.

This fix works, because it ensures the getter/setter is created on the prototype, rather than the instance. This _works_ but I’m not sure it’s the correct fix? The alternate fix is to store “instance definitions” separately from “prototype definitions”.

fixes: #11