Fix stopListening when changing a prop of type "state"
We used to just call `this.stopListening` on a property of type
"state" whenever we changed it out for some other object. However,
this also unbinds any other listeners on that object, unexpectedly.

Instead, we should keep a reference to the handler, which we do by
changing `_getEventBubblingHandler` into
`_getCachedEventBubblingHandler` and saving it for the `propertyName`.
This way, we can unbind just the handler we just set, and leave other
event listeners untouched.