Collections model props as deps in derived.
Allows specifying collection's model property as the dependency in derived property.  So if given property will change in any of the collection's models, derived will be re-calculated. Introduces private property _collectionEvents, to store already subscribed events, and to do not subscribe twice for the same event in case if multiple derived properties will use the same dependency.